:art: let's generalize the code so that we don't have to copy paste same is multiple places, also it's fun when you write the code and it just works on first run :)
